Revert "- 생성 과 업데이트 dto 분리 (#260)"
This reverts commit 269588c8c09ee670c5947aefffed9562021841f3.

+++ src/main/java/com/patientpal/backend/caregiver/dto/response/CaregiverRankingResponse.java
@@ -12,10 +12,10 @@ public class CaregiverRankingResponse {
     private Long id;
     private String name;
     private String address;
-    private float rating;
+    private double rating;
 
     @Builder
-    public CaregiverRankingResponse(Long id, String name, String address, float rating) {
+    public CaregiverRankingResponse(Long id, String name, String address, double rating) {
         this.id = id;
         this.name = name;
         this.address = address;
